1/5 Tests created and passed
Tests involving the method checkAccusation() are complete, and code functionality has been added so that these tests pass. The method itsefl takes in strings as arguments, as well as needs the solution object to bebale to work. Also refactored Card class to be able to pass in CardType enums into constructor, and refactored Solution to have objects instead of being static.