Finished C21A-2
Code is in the file ClueCardsPanel. Specifications met. The card GUI has all relevant components. Testing was done within the file that a) insantiated a player and shown his cards in the in hand field, and b) seen cards from other players have also been tested and added, with full color coding.