Started C22A
Implemented basic functionality. PaintComponent() works with draw() to create a graphic of the board. Only walkways, empty space and rooms are completed. Rooms are only generic so far; need to be further developed. Size of the board also needs to be instantiated. Rooms need to be further modified so as to show door direction. Need to add a title. Need to add players. Finally, will need to be able to be displayed along with two other panels made in previous assignments.